[project @ 2002-07-08 08:12:02 by simonpj]
-------------------------------------
	Add Template Haskell syntax data type
	-------------------------------------

Ideally this data type should be the same as the existing
HsSource, but the emphasis is more on simplicity here.  There
should be some discussion on convergence, but meanwhile we
decided it was best to make progress in parallel.

darcs-hash:20020708081202-c7106-9aa8c0a6b3e423bc87a2fe42b8a66bde4308f073.gz