Move ParseUtils into exposed-modules, in haskell-src
The haskell-src package.conf.in and .cabal file were out of sync.  In
the .cabal file Language.Haskell.ParseUtils was hidden, while it was
exposed in the package.conf file.

This patch exposes ParseUtils in the .cabal file too - fixing programs
that use ParseUtils on systems (like gentoo) where the libraries are
built the cabal way.

Thanks to dcoutts for clarifying, and gour and kolmodin on #haskell for
pestering.

darcs-hash:20060912130421-9c5c1-9d5512f29299a1b9e0e647007f0cca0448380dba.gz